According to https://docs.aws.amazon.com/AmazonS3/latest/dev/UsingMetadata.html#object-metadata it is not an error that the objectstore last_modified timestamp differs from the oc_filecache mtime.
It our use of objectstore preserves nothing except filesize and contents.
We should create inode and/or directory objects along with the file contents objects.
Currently the oc_filecache is authorative for folder names, file names, permissions -- thus a tree structure
(and user ownership) cannot be reconstructed from the objectstore.